+++ src/utils/is-json.ts
@@ -4,6 +4,14 @@
  * @return Returns if the string is valid JSON
  */
 export function isJson(str: string): boolean {
+	if (!str || !str.indexOf) {
+		return false;
+	}
+
+	if (str.indexOf('{') !== 0 && str.indexOf('{') !== 0) {
+		return false;
+	}
+
 	try {
 		JSON.parse(str);
 	} catch (e) {