+++ test/spec/chat/server.spec.ts
@@ -13,7 +13,7 @@ let serverfork;
 beforeAll(async () => {
 	jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000;
 
-	serverfork = await forkServer('./lib/test/examples/chat/server.js');
+	serverfork = await forkServer('./dist/test/examples/chat/server.js');
 
 	// Database
 	await pointy.db